Move AppHeader1 component to elements dir
