+++ pages/index.vue
@@ -82,7 +82,7 @@ import links from '~/modules/links.json'
 import skills from '~/modules/skills.json'
 import AppBooks from '~/components/AppBooks'
 import AppGifteeButton from '~/components/AppGifteeButton'
-import AppHeader1 from '~/components/AppHeader1'
+import AppHeader1 from '~/components/elements/AppHeader1'
 import AppHeader2 from '~/components/AppHeader2'
 import AppLinkButton from '~/components/elements/AppLinkButton'
 import AppPayPayButton from '~/components/AppPayPayButton'
